move multi test to another class
